backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg>=0.30.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
boto3==1.34.0
celery==5.3.4
redis==5.0.1
pytesseract==0.3.10
pillow>=10.2.0
pdf2image==1.16.3
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
